More bipropellant work  - Fixed a bug before it could happen with rate being an int  - It compiles now!

+++ src/main/java/zmaster587/advancedRocketry/tile/TileRocketBuilder.java
@@ -832,7 +832,7 @@ public class TileRocketBuilder extends TileEntityRFConsumer implements IButtonIn
 	public int getData(int id) {
 		switch(id) {
 		case 0:
-			return getRocketStats().getFuelRate(FuelType.LIQUID_MONOPROPELLANT);
+			return (int)getRocketStats().getFuelRate(FuelType.LIQUID_MONOPROPELLANT);
 		case 1:
 			return getRocketStats().getWeight();
 		case 2:
